backend/requirements.txt
Django==4.2.1
djangorestframework==3.14.0
django-cors-headers==3.14.0
dj-database-url==0.5.0
gunicorn==20.1.0
psycopg2>=2.8
sqlparse==0.4.2
